fix(renaming): rename folder
